changing file name and separating nominatim queries
